+++ src/MudBlazor.Docs/Pages/Components/DropZone/DropZonePage.razor
@@ -27,7 +27,7 @@
 				<DocsPageSection>
 					<SectionHeader Title="Dragging style">
 						<Description>
-							Not Working Benno, see discord.
+							
 						</Description>
 					</SectionHeader>
 					<SectionContent ShowCode="false" Code="DropZoneDraggingStyleExample">
@@ -37,7 +37,7 @@
 				<DocsPageSection>
 					<SectionHeader Title="Can drop styles">
 						<Description>
-							Not Working Benno, see discord.
+							
 						</Description>
 					</SectionHeader>
 					<SectionContent ShowCode="false" Code="DropZoneCanDropStylesExample">
@@ -50,7 +50,7 @@
 		<DocsPageSection>
 			<SectionHeader Title="Dragging style">
 				<Description>
-					Not Working Benno, see discord.
+					
 				</Description>
 			</SectionHeader>
 			<SectionContent ShowCode="false" Code="DropZoneDraggingStyleExample">
@@ -61,7 +61,7 @@
 		<DocsPageSection>
 			<SectionHeader Title="Disabled items">
 				<Description>
-					Should work Jonny, see discord.
+					
 				</Description>
 			</SectionHeader>
 			<SectionContent ShowCode="false" Code="DropZoneDisabledExample">
@@ -69,18 +69,7 @@
 			</SectionContent>
 		</DocsPageSection>
 
-		<DocsPageSection>
-			<SectionHeader Title="DEV">
-				<Description>
-
-				</Description>
-			</SectionHeader>
-			<SectionContent DarkenBackground="true" Code="DropZoneAdvancedExample">
-				<DropZoneAdvancedExample />
-			</SectionContent>
-		</DocsPageSection>
-
-		<DocsPageSection>
+        <DocsPageSection>
 			<SectionHeader Title="Simple Kanban Board">
 				<Description>
