+++ src/MudBlazor.Docs/Pages/Components/DropZone/Examples/DropZoneDisabledExample.razor
@@ -1,20 +1,20 @@
 ﻿@namespace MudBlazor.Docs.Examples
 
-<MudItemDropContainer T="FileItem" ItemIsDisbaled="@( (item) => item.IsReadOnly == true)"  Items="_items" ItemsSelector="@((item,dropzone) => item.TransferSlot == dropzone)"  ItemDropped="ItemUpdated" Class="d-flex">
+<MudDropContainer T="FileItem" ItemIsDisbaled="@( (item) => item.IsReadOnly == true)"  Items="_items" ItemsSelector="@((item,dropzone) => item.TransferSlot == dropzone)"  ItemDropped="ItemUpdated" Class="d-flex">
 	<ChildContent>
-		<MudItemDropZone T="FileItem" Identifier="My" Class="mx-8 pa-4 docs-gray-bg rounded">
+		<MudDropZone T="FileItem" Identifier="My" Class="mx-8 pa-4 docs-gray-bg rounded">
 			<MudText Typo="Typo.h6" Class="mb-4">My files</MudText>
-		</MudItemDropZone>
-		<MudItemDropZone T="FileItem" Identifier="Remote" Class="mx-8 pa-4 docs-gray-bg rounded">
+		</MudDropZone>
+		<MudDropZone T="FileItem" Identifier="Remote" Class="mx-8 pa-4 docs-gray-bg rounded">
 			<MudText Typo="Typo.h6" Class="mb-4">Remote Files</MudText>
-		</MudItemDropZone>
+		</MudDropZone>
 	</ChildContent>
 	<ItemRenderer>
 		<MudPaper Elevation="25" Class="pa-4 my-2">
 			<MudText Color="@(context.IsReadOnly? Color.Success : Color.Default)">@context.Name</MudText>
 		</MudPaper>
 	</ItemRenderer>
-</MudItemDropContainer>
+</MudDropContainer>
 
 
 @code {
