+++ src/MudBlazor.Docs/Pages/Components/DropZone/Examples/DropZoneDraggingStyleExample.razor
@@ -1,18 +1,18 @@
 ﻿@namespace MudBlazor.Docs.Examples
 
-<MudItemDropContainer T="DropItem" Items="_items" DraggingClass="mud-theme-success" ItemDraggingClass="mud-theme-info" ItemsSelector="@((item,dropzone) => item.Identifier == dropzone)" ItemDropped="ItemUpdated" Class="d-flex">
+<MudDropContainer T="DropItem" Items="_items" DraggingClass="mud-theme-success" ItemDraggingClass="mud-theme-info" ItemsSelector="@((item,dropzone) => item.Identifier == dropzone)" ItemDropped="ItemUpdated" Class="d-flex">
     <ChildContent>
-        <MudItemDropZone T="DropItem" Identifier="Drop Zone 1"  Class="mx-8 pa-4 docs-gray-bg rounded">
+        <MudDropZone T="DropItem" Identifier="Drop Zone 1"  Class="mx-8 pa-4 docs-gray-bg rounded">
             <MudText Typo="Typo.h6" Class="mb-4">Drop Zone 1</MudText>
-        </MudItemDropZone>
-        <MudItemDropZone T="DropItem" Identifier="Drop Zone 2" DraggingClass2="mud-theme-warning" Class="mx-8 pa-4 docs-gray-bg rounded">
+        </MudDropZone>
+        <MudDropZone T="DropItem" Identifier="Drop Zone 2" DraggingClass2="mud-theme-warning" Class="mx-8 pa-4 docs-gray-bg rounded">
             <MudText Typo="Typo.h6" Class="mb-4">Drop Zone 2</MudText>
-        </MudItemDropZone>
+        </MudDropZone>
     </ChildContent>
     <ItemRenderer>
         <MudPaper Elevation="25" Class="pa-4 my-2">@context.Name</MudPaper>
     </ItemRenderer>
-</MudItemDropContainer>
+</MudDropContainer>
 
 @code {
     private void ItemUpdated(MudItemDropInfo<DropItem> dropItem)
