+++ src/MudBlazor.Docs/Services/MenuService.cs
@@ -37,8 +37,8 @@ namespace MudBlazor.Docs.Services
             .AddItem("ChipSet", typeof(MudChipSet))
             .AddItem("Badge", typeof(MudBadge))
             .AddItem("AppBar", typeof(MudAppBar))
-            .AddItem("Drop Zone", typeof(MudDrawer), typeof(MudDrawerHeader), typeof(MudDrawerContainer))
-            .AddItem("Drawer", typeof(MudDropZone<T>), typeof(MudDropItem<T>))
+            .AddItem("Drawer", typeof(MudDrawer), typeof(MudDrawerHeader), typeof(MudDrawerContainer))
+            .AddItem("Drop Zone", typeof(MudDropZone<T>), typeof(MudDropContainer<T>), typeof(MudDynamicDropItem<T>))
             .AddItem("Link", typeof(MudLink))
             .AddItem("Menu", typeof(MudMenu), typeof(MudMenuItem))
             .AddItem("Message Box", typeof(MudMessageBox))
