+++ src/MudBlazor/Components/DropZone/MudDropContainer.razor
@@ -8,3 +8,7 @@
 		@ChildContent
 	</div>
 </CascadingValue>
+
+@code {
+
+}
