--- src/MudBlazor/Components/DropZone/MudDropItem.razor
@@ -1,12 +1,0 @@
-﻿@namespace MudBlazor
-@inherits MudComponentBase
-@typeparam T
-
-<div 
-	class="@Classname" style="@Style"
-	draggable="@(Disable == false ? "true" : "false")"
-    @ondragstart="DragStarted"
-	@ondragend="DragEnded"
-	@attributes="UserAttributes">
-    @ChildContent
-</div>
