+++ src/MudBlazor/Components/DropZone/MudDropZone.razor
@@ -2,12 +2,36 @@
 @inherits MudComponentBase
 @typeparam T
 
-<div 
-	class="@Classname" style="@Style" 
-	ondragover="event.preventDefault();" ondragstart="event.dataTransfer.setData('', event.target.id);"
-        @ondrop="HandleDrop"
-        @ondragenter="HandleDragEnter"
-        @ondragleave="HandleDragLeave"
-@attributes="UserAttributes">
-    @ChildContent
+<div class="@Classname" style="@Style"
+	 ondragover="event.preventDefault();" ondragstart="event.dataTransfer.setData('', event.target.id);"
+	 @ondrop="HandleDrop"
+	 @ondragenter="HandleDragEnter"
+	 @ondragleave="HandleDragLeave"
+	 @attributes="UserAttributes">
+
+	@ChildContent
+	
+	@foreach (var item in GetItems())
+	{
+		<MudDynamicDropItem 
+			@key="@(item.GetHashCode())" T="T" DraggingClass="@GetItemDraggingClass()" 
+			Item="item" HideOnDrag="HideItemOnDrag" 
+			OnDragStarted="DragOperationStarted" 
+			OnDragEnded="FinishedDragOperation"
+			ZoneIdentifier="@Identifier"
+			Disabled="@GetItemDisabledStatus(item)"
+			DisabledClass=@(DisabledClass ?? Container?.DisabledClass)
+		>
+			@{
+				var renderer = GetItemTemplate();
+			}
+
+			@renderer(item)
+
+		</MudDynamicDropItem>
+	}
 </div>
+
+@code {
+
+}
