--- src/MudBlazor/Components/DropZone/MudItemDropContainer.razor
@@ -1,14 +1,0 @@
-﻿@namespace MudBlazor
-@inherits MudComponentBase
-@typeparam T
-
-<CascadingValue Value="this" IsFixed="true">
-	<div style="@Style"
-		 class="@Classname">
-		@ChildContent
-	</div>
-</CascadingValue>
-
-@code {
-
-}
